+++ crates/xtask-unpublished/src/main.rs
@@ -1,4 +1,3 @@
-mod unpublished;
 mod xtask;
 
 fn main() {
