--- crates/xtask/src/xtask.rs
@@ -1,83 +1,0 @@
-use cargo::util::command_prelude::*;
-
-pub fn cli() -> clap::Command {
-    clap::Command::new("xtask")
-        .subcommand_required(true)
-        .arg_required_else_help(true)
-        .arg(
-            opt(
-                "verbose",
-                "Use verbose output (-vv very verbose/build.rs output)",
-            )
-            .short('v')
-            .action(ArgAction::Count)
-            .global(true),
-        )
-        .arg_quiet()
-        .arg(
-            opt("color", "Coloring: auto, always, never")
-                .value_name("WHEN")
-                .global(true),
-        )
-        .arg(flag("frozen", "Require Cargo.lock and cache are up to date").global(true))
-        .arg(flag("locked", "Require Cargo.lock is up to date").global(true))
-        .arg(flag("offline", "Run without accessing the network").global(true))
-        .arg(multi_opt("config", "KEY=VALUE", "Override a configuration value").global(true))
-        .arg(
-            Arg::new("unstable-features")
-                .help("Unstable (nightly-only) flags to Cargo, see 'cargo -Z help' for details")
-                .short('Z')
-                .value_name("FLAG")
-                .action(ArgAction::Append)
-                .global(true),
-        )
-        .subcommands([crate::unpublished::cli()])
-}
-
-pub fn exec(args: &clap::ArgMatches, config: &mut cargo::util::Config) -> cargo::CliResult {
-    config_configure(config, args)?;
-
-    match args.subcommand() {
-        Some(("unpublished", args)) => crate::unpublished::exec(args, config)?,
-        Some((name, _)) => unreachable!("clap enforces {name} to not exist"),
-        None => unreachable!("clap enforces a subcommand is present"),
-    }
-
-    Ok(())
-}
-
-fn config_configure(config: &mut Config, args: &ArgMatches) -> CliResult {
-    let verbose = args.verbose();
-    // quiet is unusual because it is redefined in some subcommands in order
-    // to provide custom help text.
-    let quiet = args.flag("quiet");
-    let color = args.get_one::<String>("color").map(String::as_str);
-    let frozen = args.flag("frozen");
-    let locked = args.flag("locked");
-    let offline = args.flag("offline");
-    let mut unstable_flags = vec![];
-    if let Some(values) = args.get_many::<String>("unstable-features") {
-        unstable_flags.extend(values.cloned());
-    }
-    let mut config_args = vec![];
-    if let Some(values) = args.get_many::<String>("config") {
-        config_args.extend(values.cloned());
-    }
-    config.configure(
-        verbose,
-        quiet,
-        color,
-        frozen,
-        locked,
-        offline,
-        &None,
-        &unstable_flags,
-        &config_args,
-    )?;
-    Ok(())
-}
-
-#[test]
-fn verify_cli() {
-    cli().debug_assert();
-}
